Make NodeHealthCheckerService as pluggable class in NodeManager
