+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-nodemanager/src/main/java/org/apache/hadoop/yarn/server/nodemanager/health/NodeHealthCheckerService.java
@@ -18,102 +18,17 @@
 
 package org.apache.hadoop.yarn.server.nodemanager.health;
 
-import org.apache.hadoop.thirdparty.com.google.common.annotations.VisibleForTesting;
-import org.apache.hadoop.thirdparty.com.google.common.base.Joiner;
-import org.apache.hadoop.thirdparty.com.google.common.base.Strings;
-import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.service.CompositeService;
-import org.apache.hadoop.service.Service;
-import org.apache.hadoop.yarn.conf.YarnConfiguration;
 import org.apache.hadoop.yarn.server.nodemanager.LocalDirsHandlerService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 /**
- * This class provides functionality of checking the health of a node and
- * reporting back to the service for which the health checker has been asked to
- * report.
- *
- * It is a {@link CompositeService}: every {@link Service} must be registered
- * first in serviceInit, and should also implement the {@link HealthReporter}
- * interface - otherwise an exception is thrown.
- *
- * Calling functions of HealthReporter merge its dependent
- * services' reports.
- *
- * @see HealthReporter
- * @see LocalDirsHandlerService
- * @see TimedHealthReporterService
+ * This is the base class for NodeHealthCheckerService.
+ * The default implementation is {@link NodeHealthCheckerServiceImpl}
  */
-public class NodeHealthCheckerService extends CompositeService
-    implements HealthReporter {
+public abstract class NodeHealthCheckerService extends CompositeService {
 
-  public static final Logger LOG =
-      LoggerFactory.getLogger(NodeHealthCheckerService.class);
-  private static final int MAX_SCRIPTS = 4;
-
-  private List<HealthReporter> reporters;
-  private LocalDirsHandlerService dirsHandler;
-  private ExceptionReporter exceptionReporter;
-
-  public static final String SEPARATOR = ";";
-
-  public NodeHealthCheckerService(
-      LocalDirsHandlerService dirHandlerService) {
-    super(NodeHealthCheckerService.class.getName());
-
-    this.reporters = new ArrayList<>();
-    this.dirsHandler = dirHandlerService;
-    this.exceptionReporter = new ExceptionReporter();
-  }
-
-  @Override
-  protected void serviceInit(Configuration conf) throws Exception {
-    reporters.add(exceptionReporter);
-    addHealthReporter(dirsHandler);
-    String[] configuredScripts = conf.getTrimmedStrings(
-        YarnConfiguration.NM_HEALTH_CHECK_SCRIPTS,
-        YarnConfiguration.DEFAULT_NM_HEALTH_CHECK_SCRIPTS);
-    if (configuredScripts.length > MAX_SCRIPTS) {
-      throw new IllegalArgumentException("Due to performance reasons " +
-          "running more than " + MAX_SCRIPTS + "scripts is not allowed.");
-    }
-    for (String configuredScript : configuredScripts) {
-      addHealthReporter(NodeHealthScriptRunner.newInstance(
-          configuredScript, conf));
-    }
-    super.serviceInit(conf);
-  }
-
-  /**
-   * Adds a {@link Service} implementing the {@link HealthReporter} interface,
-   * if that service has not been added to this {@link CompositeService} yet.
-   *
-   * @param service to add
-   * @throws Exception if not a {@link HealthReporter}
-   *         implementation is provided to this function
-   */
-  @VisibleForTesting
-  void addHealthReporter(Service service) throws Exception {
-    if (service != null) {
-      if (getServices().stream()
-          .noneMatch(x -> x.getName().equals(service.getName()))) {
-        if (!(service instanceof HealthReporter)) {
-          throw new Exception("Attempted to add service to " +
-              "NodeHealthCheckerService that does not implement " +
-              "HealthReporter.");
-        }
-        reporters.add((HealthReporter) service);
-        addService(service);
-      } else {
-        LOG.debug("Omitting duplicate service: {}.", service.getName());
-      }
-    }
+  public NodeHealthCheckerService(String name) {
+    super(name);
   }
 
   /**
@@ -121,44 +36,26 @@ public class NodeHealthCheckerService extends CompositeService
    *
    * @return the report string about the health of the node
    */
-  @Override
-  public String getHealthReport() {
-    ArrayList<String> reports = reporters.stream()
-        .map(reporter -> Strings.emptyToNull(reporter.getHealthReport()))
-        .collect(Collectors.toCollection(ArrayList::new));
-    return Joiner.on(SEPARATOR).skipNulls().join(reports);
-  }
+  public abstract String getHealthReport();
 
   /**
    * @return <em>true</em> if the node is healthy
    */
-  @Override
-  public boolean isHealthy() {
-    return reporters.stream().allMatch(HealthReporter::isHealthy);
-  }
+  public abstract boolean isHealthy();
 
   /**
    * @return when the last time the node health status is reported
    */
-  @Override
-  public long getLastHealthReportTime() {
-    Optional<Long> max = reporters.stream()
-        .map(HealthReporter::getLastHealthReportTime).max(Long::compareTo);
-    return max.orElse(0L);
-  }
+  public abstract long getLastHealthReportTime();
 
   /**
    * @return the disk handler
    */
-  public LocalDirsHandlerService getDiskHandler() {
-    return dirsHandler;
-  }
+  public abstract LocalDirsHandlerService getDiskHandler();
 
   /**
    * Propagating an exception to {@link ExceptionReporter}.
    * @param exception the exception to propagate
    */
-  public void reportException(Exception exception) {
-    exceptionReporter.reportException(exception);
-  }
+  public abstract void reportException(Exception exception);
 }
