+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-nodemanager/src/test/java/org/apache/hadoop/yarn/server/nodemanager/containermanager/BaseContainerManagerTest.java
@@ -27,6 +27,7 @@ import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.doNothing;
 
 import org.apache.hadoop.yarn.server.nodemanager.NodeResourceMonitorImpl;
+import org.apache.hadoop.yarn.server.nodemanager.health.NodeHealthCheckerServiceImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -172,6 +173,12 @@ public abstract class BaseContainerManagerTest {
     this.nodeStatusUpdater = nodeStatusUpdater;
   }
 
+  public void setNodeHealthCheckerService(NodeHealthCheckerService nhcs,
+      Configuration yarnConfiguration) {
+    this.nodeHealthCheckerService = nhcs;
+    this.nodeHealthCheckerService.init(yarnConfiguration);
+  }
+
   protected ContainerExecutor createContainerExecutor() {
     DefaultContainerExecutor exec = new DefaultContainerExecutor();
     exec.setConf(conf);
@@ -208,7 +215,7 @@ public abstract class BaseContainerManagerTest {
 
     dirsHandler = new LocalDirsHandlerService();
     dirsHandler.init(conf);
-    nodeHealthCheckerService = new NodeHealthCheckerService(dirsHandler);
+    nodeHealthCheckerService = new NodeHealthCheckerServiceImpl(dirsHandler);
     nodeStatusUpdater = new NodeStatusUpdaterImpl(
         context, new AsyncDispatcher(), nodeHealthCheckerService, metrics) {
       @Override
